Beautified the debug draw lines

+++ src/engine/core/Debug.h
@@ -14,12 +14,12 @@ namespace debug {
         Vector2 start;
         Vector2 end;
         Color color;
-        int thickness = 1;
+        float thickness = 1.0f;
     };
 
     void render_debug_draws(SDL_Renderer* renderer, const CameraComponent* camera);
 
-    void draw_line(const Vector2& start, const Vector2& end, const Color& color, int thickness = 1);
+    void draw_line(const Vector2& start, const Vector2& end, const Color& color, float thickness = 1.0f);
 }
